+++ nav_algo/event_helper/search.py
@@ -1,4 +1,4 @@
-from navigation_helper import *
+#from navigation_helper import *
 from navigation import *
 # TODO: figure out what else to import
